#339: Move FilePathStyleConverter to utils
