+++ hydrolib/core/basemodel.py
@@ -32,6 +32,7 @@ from pydantic.fields import ModelField, PrivateAttr
 
 from hydrolib.core.base import DummmyParser, DummySerializer
 from hydrolib.core.utils import (
+    FilePathStyleConverter,
     OperatingSystem,
     PathStyle,
     get_operating_system,
@@ -260,107 +261,6 @@ class ResolveRelativeMode(IntEnum):
 
     ToParent = 0
     ToAnchor = 1
-
-
-class FilePathStyleConverter:
-    """Class for converting file paths between different path styles."""
-
-    def __init__(self) -> None:
-        """Initializes a new instance of the FilePathStyleResolver class."""
-        self._os_path_style = get_path_style_for_current_operating_system()
-
-    def convert_to_os_style(self, file_path: Path, source_path_style: PathStyle) -> str:
-        """Resolve the file path by converting it from its own file path style to the path style for the current operating system.
-
-        Args:
-            file_path (Path): The file path to convert to the OS path style.
-            file_path_style (PathStyle): The file path style of the given file path.
-
-        Returns:
-            str: The converted file path with OS path style.
-
-        Raises:
-            NotImplementedError: When this function is called with a PathStyle other than WINDOWSLIKE or UNIXLIKE.
-        """
-
-        return FilePathStyleConverter._convert(
-            file_path, source_path_style, self._os_path_style
-        )
-
-    def convert_from_os_style(
-        self, file_path: Path, target_path_style: PathStyle
-    ) -> str:
-        """Resolve the file path by converting it from the path style for the current operating system to the target path style.
-
-        Args:
-            file_path (Path): The file path to convert to the OS path style.
-            target_path_style (PathStyle): The target file path style to which the file path should be converted.
-
-        Returns:
-            str: The converted file path with the target path style.
-
-        Raises:
-            NotImplementedError: When this function is called with a PathStyle other than WINDOWSLIKE or UNIXLIKE.
-        """
-
-        return FilePathStyleConverter._convert(
-            file_path, self._os_path_style, target_path_style
-        )
-
-    @classmethod
-    def _convert(
-        cls, file_path: Path, source_path_style: PathStyle, target_path_style: PathStyle
-    ) -> str:
-        if source_path_style == target_path_style:
-            return str(file_path)
-
-        if (
-            source_path_style == PathStyle.UNIXLIKE
-            and target_path_style == PathStyle.WINDOWSLIKE
-        ):
-            return FilePathStyleConverter._from_posix_to_windows_path(file_path)
-        elif (
-            source_path_style == PathStyle.WINDOWSLIKE
-            and target_path_style == PathStyle.UNIXLIKE
-        ):
-            return FilePathStyleConverter._from_windows_to_posix_path(file_path)
-        else:
-            raise NotImplementedError(
-                f"Cannot convert {source_path_style} to {target_path_style}"
-            )
-
-    @classmethod
-    def _from_posix_to_windows_path(cls, posix_path: Path) -> str:
-        is_relative = not posix_path.as_posix().startswith("/")
-
-        if is_relative:
-            return posix_path.as_posix()
-
-        root = posix_path.parts[1]
-        windows_root = root + ":/"
-        parts = posix_path.parts[2:]
-
-        windows_path = windows_root + "/".join(parts)
-
-        return windows_path
-
-    @classmethod
-    def _from_windows_to_posix_path(cls, windows_path: Path) -> str:
-        windows_path_str = str(windows_path).replace("\\", "/")
-        windows_path = Path(windows_path_str)
-
-        root = windows_path.parts[0]
-        is_relative = ":" not in root
-
-        if is_relative:
-            return windows_path_str
-
-        posix_root = "/" + root.split(":")[0] + "/"
-        parts = windows_path.parts[1:]
-
-        posix_path = posix_root + "/".join(parts)
-
-        return posix_path
 
 
 class FileCasingResolver:
