+++ tests/test_basemodel.py
@@ -13,7 +13,6 @@ from hydrolib.core.basemodel import (
     FileModel,
     FileModelCache,
     FilePathResolver,
-    FilePathStyleConverter,
     ModelLoadSettings,
     ModelSaveSettings,
     ParsableFileModel,
@@ -33,11 +32,6 @@ _external_path = test_output_dir / "test_save_and_load_maintains_correct_paths_e
 def runs_from_docker() -> bool:
     """Check to see if we are running from within docker."""
     return Path("/.dockerenv").exists()
-
-
-def runs_on_windows() -> bool:
-    """Check to see if we are running on Windows."""
-    return platform.system() == "Windows"
 
 
 class TestFileModel:
@@ -789,137 +783,3 @@ class TestSerializerConfig:
     def test_default(self):
         config = SerializerConfig()
         assert config.float_format == ""
-
-
-class TestFilePathStyleConverter:
-    @pytest.mark.skipif(
-        not runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    def test_convert_to_os_style_absolute_unixlike_filepath(self):
-        unix_path = "/c/path/to.file"
-        converter = FilePathStyleConverter()
-        windows_path = converter.convert_to_os_style(
-            Path(unix_path), PathStyle.UNIXLIKE
-        )
-
-        assert windows_path == "c:/path/to.file"
-
-    @pytest.mark.skipif(
-        not runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    def test_convert_to_os_style_relative_unixlike_filepath(self):
-        unix_path = "path/to.file"
-        converter = FilePathStyleConverter()
-        windows_path = converter.convert_to_os_style(
-            Path(unix_path), PathStyle.UNIXLIKE
-        )
-
-        assert windows_path == "path/to.file"
-
-    @pytest.mark.skipif(
-        runs_on_windows(),
-        reason="Platform dependent test: should only succeed on non-Windows OS.",
-    )
-    @pytest.mark.parametrize(
-        "windows_path",
-        [
-            pytest.param("c:\\path\\to.file", id="Backward slashes"),
-            pytest.param("c:/path/to.file", id="Forward slashes"),
-        ],
-    )
-    def test_convert_to_os_style_absolute_windowslike_filepath(self, windows_path: str):
-        converter = FilePathStyleConverter()
-        unix_path = converter.convert_to_os_style(
-            Path(windows_path), PathStyle.WINDOWSLIKE
-        )
-
-        assert unix_path == "/c/path/to.file"
-
-    @pytest.mark.skipif(
-        runs_on_windows(),
-        reason="Platform dependent test: should only succeed on non-Windows OS.",
-    )
-    @pytest.mark.parametrize(
-        "windows_path",
-        [
-            pytest.param("path\\to.file", id="Backward slashes"),
-            pytest.param("path/to.file", id="Forward slashes"),
-        ],
-    )
-    def test_convert_to_os_style_relative_windowslike_filepath(self, windows_path: str):
-        converter = FilePathStyleConverter()
-        unix_path = converter.convert_to_os_style(
-            Path(windows_path), PathStyle.WINDOWSLIKE
-        )
-
-        assert unix_path == "path/to.file"
-
-    @pytest.mark.skipif(
-        not runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    @pytest.mark.parametrize(
-        "windows_path",
-        [
-            pytest.param("c:\\path\\to.file", id="Backward slashes"),
-            pytest.param("c:/path/to.file", id="Forward slashes"),
-        ],
-    )
-    def test_convert_from_os_style_to_absolute_unixlike_filepath(
-        self, windows_path: str
-    ):
-        converter = FilePathStyleConverter()
-        unix_path = converter.convert_from_os_style(
-            Path(windows_path), PathStyle.UNIXLIKE
-        )
-
-        assert unix_path == "/c/path/to.file"
-
-    @pytest.mark.skipif(
-        not runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    @pytest.mark.parametrize(
-        "windows_path",
-        [
-            pytest.param("path\\to.file", id="Backward slashes"),
-            pytest.param("path/to.file", id="Forward slashes"),
-        ],
-    )
-    def test_convert_from_os_style_to_relative_unixlike_filepath(
-        self, windows_path: str
-    ):
-        converter = FilePathStyleConverter()
-        unix_path = converter.convert_from_os_style(
-            Path(windows_path), PathStyle.UNIXLIKE
-        )
-
-        assert unix_path == "path/to.file"
-
-    @pytest.mark.skipif(
-        runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    def test_convert_from_os_style_to_absolute_windowslike_filepath(self):
-        unix_path = "/c/path/to.file"
-        converter = FilePathStyleConverter()
-        windows_path = converter.convert_from_os_style(
-            Path(unix_path), PathStyle.WINDOWSLIKE
-        )
-
-        assert windows_path == "c:/path/to.file"
-
-    @pytest.mark.skipif(
-        runs_on_windows(),
-        reason="Platform dependent test: should only succeed on Windows OS.",
-    )
-    def test_convert_from_os_style_to_relative_windowslike_filepath(self):
-        unix_path = "path/to.file"
-        converter = FilePathStyleConverter()
-        windows_path = converter.convert_from_os_style(
-            Path(unix_path), PathStyle.WINDOWSLIKE
-        )
-
-        assert windows_path == "path/to.file"
